requirements.txt
args==0.1.0
boto==2.34.0
clint==0.4.1
Flask==0.10.1
gnureadline==6.3.3
google-api-python-client==1.3.1
httplib2==0.9
itsdangerous==0.24
Jinja2==2.7.3
keyring==4.0
Markdown==2.4.1
MarkupSafe==0.23
metalsmyth==0.1.3
oauth2client==1.4.2
pyasn1-modules==0.0.5
pyasn1==0.1.7
python-dateutil==2.3
python-frontmatter==0.1.5
PyYAML==3.11
requests==2.3.0
rsa==3.1.4
sh==1.09
simplejson==3.6.5
six==1.8.0
tarbell==1.0
uritemplate==0.6
Werkzeug==0.9.6
xlrd==0.9.3
